requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
pandas==1.5.2
openpyxl==3.0.10
numpy==1.24.1
geocoder==1.38.1
celery==5.2.7
beautifulsoup4==4.11.1
selenium==4.8.0
djangorestframework==3.14.0
django_filter==23.1
fuzzywuzzy == 0.18.0
django-admin-extra-buttons==1.5.6
rest_framework_simplejwt==0.0.2
translitua==1.3.2
Pillow==10.0.0
Levenshtein==0.21.1